requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==25.1.0
autoit==0.2.6
bcrypt==4.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.6
convertdate==2.4.0
cron-descriptor==1.4.5
cryptography==43.0.1
distro==1.9.0
Django==5.0.9
django-celery-beat==2.7.0
django-cron==0.6.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
keyboard==0.13.5
kombu==5.4.2
lunardate==0.2.2
MouseInfo==0.1.3
mssql-django==1.5
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.63.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
playwright==1.50.0
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.0.0
psycopg2==2.9.10
py3270==0.3.5
PyAutoGUI==0.9.54
PyAutoIt==0.6.5
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==12.1.1
PyGetWindow==0.0.9
PyJWT==2.9.0
pyluach==2.2.0
PyMeeus==0.5.12
PyMsgBox==1.0.9
pyodbc==5.1.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
schedule==1.2.2
setproctitle==1.3.4
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
waitress==3.0.0
wcwidth==0.2.13
workalendar==17.0.0
yarl==1.18.3
